Feat Sequelize config and define entities relations

+++ src/module/products/products.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateProductDTO } from './dto/create.product.dto';
 import { UpdateProductDTO } from './dto/update.product.dto';
 import { ProductsRepository } from './repositories/products.repository';
@@ -9,6 +9,7 @@ export class ProductsService {
 
   async create(createProductDTO: CreateProductDTO) {
     const product = await this.productRepository.create(createProductDTO);
+
     return product;
   }
 
@@ -18,16 +19,33 @@ export class ProductsService {
 
   async findOne(id: string) {
     const product = await this.productRepository.findOne(id);
+
+    if (!product) {
+      throw new NotFoundException('Product not found.');
+    }
+
     return product;
   }
 
   async update(id: string, updateProductDTO: UpdateProductDTO) {
     const product = await this.productRepository.update(id, updateProductDTO);
+
+    if (!product) {
+      throw new NotFoundException('Product not found.');
+    }
+
     return product;
   }
 
   async remove(id: string) {
+    const product = await this.productRepository.findOne(id);
+
+    if (!product) {
+      throw new NotFoundException('Product not found.');
+    }
+
     await this.productRepository.delete(id);
+
     return;
   }
 }
